Added collision and the ship will now stop when it hits a planet
