+++ Assets/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: ecbcb05b01f99824e872a476dbec93e8
+folderAsset: yes
+timeCreated: 1492895449
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
