+++ Assets/Scripts/BoxCollideController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 434bb64beabe8144e91c68c26f75285b
+timeCreated: 1492897499
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
